Refactor Forum Events Plugin
Made it so that the course coach pulls from this plugin rather than relying on the data pulled through the other field in the event record. This allows us to centralise this funcitonality and takes the dependancy out of the events. It also allows us to make use of core moodle events not just our custom ones as it just uses the courseid to determine the course coach.